refactor: splits core.py into stage-based modules
